# packages/shared-types/src/index.ts
export type UserRole = 'super_admin' | 'section_admin' | 'coach' | 'parent';

export type PaymentMethod = 'auto_link' | 'auto_qr' | 'manual_transfer';

export type AttendanceStatus = 'expected' | 'present' | 'late' | 'absent';